pyproject.toml
[project]
name = "lecture-live-translator"
version = "0.1.0"
description = "Real-time lecture transcription and translation with Google Cloud Speech-to-Text and Translation API."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-cloud-speech>=2.33.0",
    "google-cloud-storage>=3.4.0",
    "google-cloud-translate>=3.20.4",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "ruff>=0.13.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
